Creates Error Token on integers that are out of range
In order to generate correct output, the scanner will create an error token, when it tries to read an integer that cannot be converted to a long. Fixes #5

+++ src/main/Token/IntegerToken.cpp
@@ -23,6 +23,6 @@ long long IntegerToken::getValue() {
     return value;
 }
 
-void IntegerToken::setValue(long long value) {
+void IntegerToken::setValue(long value) {
     this->value = value;
 }
